requirements.txt
flask
flask-cors
feedparser
vaderSentiment
requests_cache
python-dotenv
transformers
torch
lxml
supabase
requests
yfinance
pandas
scikit-learn
pydantic
xgboost
